[jscoq] [modules] Refactor a some classes.
Put more classes on their own modules so future patches/diffs are clearer.
